Secure Chat Application with End-to/requirements.txt
flask
flask-socketio
cryptography
pycryptodome
